"Replaced standard RichTextBlock with custom RichTextBlockWithFeatures"
