+++ home/blocks/TabsBlock.py
@@ -1,5 +1,6 @@
 from wagtail.core import blocks
-from wagtail.core.blocks.field_block import RichTextBlock, RawHTMLBlock, CharBlock, BooleanBlock
+from wagtail.core.blocks.field_block import RawHTMLBlock, CharBlock, BooleanBlock
+from RichTextBlockWithFeatures import RichTextBlock
 from wagtail.core.blocks.stream_block import StreamBlock
 from wagtail.images.blocks import ImageChooserBlock
